move timestamp code into a utility function
